1.제품 서비스 테스트 코드 구현

+++ src/main/java/TeamCamp/demo/service/ProductService.java
@@ -3,8 +3,6 @@ package TeamCamp.demo.service;
 import TeamCamp.demo.common.s3.AwsS3Service;
 import TeamCamp.demo.common.s3.FileService;
 import lombok.RequiredArgsConstructor;
-import net.sf.ehcache.util.ProductInfo;
-import org.apache.commons.compress.utils.FileNameUtils;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,9 +15,10 @@ import TeamCamp.demo.exception.product.ProductNotFoundException;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
-import java.util.Optional;
+import java.util.List;
+import java.util.stream.Collectors;
 
-import static TeamCamp.demo.util.FilePathConstants.PRODUCT_IMAGES_DIR;
+import static TeamCamp.demo.dto.ProductDto.*;
 
 @RequiredArgsConstructor
 @Service
@@ -29,7 +28,7 @@ public class ProductService {
     private final AwsS3Service awsS3Service;
 
     @Transactional
-    public void saveProduct(ProductDto.SaveRequest request, MultipartFile productImage) throws IOException {
+    public void saveProduct(SaveRequest request, MultipartFile productImage) throws IOException {
         if (productImage != null){
             String originImagePath = awsS3Service.uploadProductImage(productImage);
             String thumbnailImagePath = FileService.toThumbnail(originImagePath);
@@ -44,23 +43,31 @@ public class ProductService {
         return productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException())
                 .toProductInfoResponse();
     }
+
+    @Cacheable(value = "product" , key = "#id")
+    public List<ProductInfoResponse> getProductInfos(){
+        return productRepository.findAll().stream()
+                .map(Product::toProductInfoResponse)
+                .collect(Collectors.toList());
+    }
     @Cacheable(value = "product",key = "#id")
     @Transactional
     public void deleteProduct(Long id){
        Product product = productRepository.findById(id)
                        .orElseThrow(ProductNotFoundException::new);
        String path = product.getOriginImagePath();
-       String key  = FileService.getFileName(path);
        productRepository.deleteById(id);
-       awsS3Service.deleteProductImage(key);
-
+        if(path != null){
+            String key  = FileService.getFileName(path);
+            awsS3Service.deleteProductImage(key);
+        }
     }
 
     @Cacheable(value = "product",key = "#id")
     @Transactional
-    public void updateProduct(Long id, ProductDto.SaveRequest updateProduct,@Nullable MultipartFile productImage){
+    public void updateProduct(Long id, SaveRequest updateProduct,@Nullable MultipartFile productImage){
         Product saveProduct = productRepository.findById(id)
-                .orElseThrow(()-> new ProductNotFoundException());
+                .orElseThrow(ProductNotFoundException::new);
         String savedImagePath = saveProduct.getOriginImagePath();
         String updateImagePath = updateProduct.getOriginImagePath();
 
@@ -78,8 +85,8 @@ public class ProductService {
     }
 
     private boolean isDeleteSavedImage(String savedImagePath, String updatedImagePath,
-                                       MultipartFile brandImage) {
+                                       MultipartFile productImage) {
         return ((updatedImagePath == null && savedImagePath != null) ||
-                (updatedImagePath != null && brandImage != null));
+                (savedImagePath != null && productImage != null));
     }
 }
